+++ src/cpu/opcodes/ldy.rs
@@ -124,7 +124,7 @@ pub fn ldy_absolute_x(cpu: &mut Cpu, bus: &mut Bus) -> u8 {
         cycles = 5;
     }
 
-    if page_crossed(base_address, target_address as u16) {
+    if page_crossed(base_address, target_address) {
         cycles += 1;
     }
